+++ src/DotNext.Tests/EnumTests.cs
@@ -8,6 +8,18 @@ namespace DotNext
     [ExcludeFromCodeCoverage]
     public sealed class EnumTests : Test
     {
+        private sealed class TestEnumValueAttribute : Attribute
+        {
+        }
+
+        private enum EnumWithAttributes
+        {
+            None = 0,
+
+            [TestEnumValue]
+            WithAttribute
+        }
+
         [Fact]
         public static void ValuesTest()
         {
@@ -76,6 +88,14 @@ namespace DotNext
         }
 
         [Fact]
+        public static void MemberOrder()
+        {
+            Equal(EnvironmentVariableTarget.Process, Enum<EnvironmentVariableTarget>.Members[0]);
+            Equal(EnvironmentVariableTarget.User, Enum<EnvironmentVariableTarget>.Members[1]);
+            Equal(EnvironmentVariableTarget.Machine, Enum<EnvironmentVariableTarget>.Members[2]);
+        }
+
+        [Fact]
         public static void HasFlag()
         {
             var flags = BindingFlags.CreateInstance | BindingFlags.Public;
@@ -123,6 +143,32 @@ namespace DotNext
             True(e.Equals(value));
             value = Enum<EnvironmentVariableTarget>.GetMember(EnvironmentVariableTarget.Process);
             False(e.Equals(value));
+        }
+
+        [Fact]
+        public static void CustomAttributeProvider()
+        {
+            ICustomAttributeProvider member = Enum<EnumWithAttributes>.GetMember(nameof(EnumWithAttributes.None));
+            False(member.IsDefined(typeof(TestEnumValueAttribute), false));
+            Empty(member.GetCustomAttributes(false));
+            Empty(member.GetCustomAttributes(typeof(TestEnumValueAttribute), false));
+
+            member = Enum<EnumWithAttributes>.GetMember(nameof(EnumWithAttributes.WithAttribute));
+            True(member.IsDefined(typeof(TestEnumValueAttribute), false));
+            NotEmpty(member.GetCustomAttributes(false));
+            NotEmpty(member.GetCustomAttributes(typeof(TestEnumValueAttribute), false));
+        }
+
+        [Fact]
+        public static void CustomAttributes()
+        {
+            var member = Enum<EnumWithAttributes>.GetMember(nameof(EnumWithAttributes.None));
+            Null(member.GetCustomAttribute<TestEnumValueAttribute>());
+            Empty(member.GetCustomAttributes<TestEnumValueAttribute>());
+
+            member = Enum<EnumWithAttributes>.GetMember(nameof(EnumWithAttributes.WithAttribute));
+            NotNull(member.GetCustomAttribute<TestEnumValueAttribute>());
+            NotEmpty(member.GetCustomAttributes<TestEnumValueAttribute>());
         }
     }
 }